pyproject.toml
[tool.poetry]
name = "llm-management"
version = "0.1.0"
description = "null"
authors = ["mySociety <alex.parsons@mysociety.org>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/mysociety/llm-management"
repository = "https://github.com/mysociety/llm-management"
include = [
    "LICENSE.md",
]

[tool.poetry_bumpversion.file."src/llm_management/__init__.py"]

[tool.poetry.scripts]
llm-management = "llm_management.__main__:app"

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
pydantic-ai = "^1.77.0"
typer = "^0.24.1"
pydantic-store = "^0.1.0"
exoscale = "^0.16.3"
pydantic-settings = "^2.13.1"
fastapi = "^0.135.3"
uvicorn = "^0.44.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pylint = "^2.12.2"
ruff = "^0.4.4"
pyright = "^1.1"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pyright]
reportAttributeAccessIssue = "none"